+++ CHANGELOG.md
@@ -1,6 +1,18 @@
-# 0.3.0
+# WIP 0.3.2
 
 * Add `DatapipeApp.api` subapp to mount extra APIs
+* Add `step status` command that shows status (number of idx's to process) for a
+  specific step
+* Add support for filtering by labels in `step` commands
+
+# 0.3.1
+
+* Add input/output tables to `datapipe step list` output
+* Add lint `LintDataWOMeta` to check and fix data without meta
+
+# 0.3.0
+
+* Add `/api/v1alpha1/labelstudio-webhook` endpoint for integration with Label Studio
 
 # 0.2.1